+++ app/src/main/java/com/task/management/workflow/task/data/remote/TaskService.kt
@@ -46,5 +46,5 @@ interface TaskService {
 
     //
     @GET("api/v1/tasks/user/{userId}")
-    suspend fun getTasksByUserId(@Path("userId") userId: Int): Response<List<TaskDto>>
+    suspend fun getTasksByUserId(@Path("userId") userId: Long): Response<List<TaskDto>>
 }
