+++ app/src/main/java/com/task/management/workflow/task/data/repository/TaskRepository.kt
@@ -17,5 +17,5 @@ class TaskRepository (
     suspend fun deleteRemotely(taskId: Long) = remoteTaskRepository.deleteRemotely(taskId)
     suspend fun getTasksByProjectAndStatusRemotely(projectId: Long, status: String?) = remoteTaskRepository.getTasksByProjectAndStatusRemotely(projectId,status)
     suspend fun getTasksByProjectAndUserAndStatusRemotely(projectId: Long, userId: Long, status: String?) = remoteTaskRepository.getTasksByProjectAndUserAndStatusRemotely(projectId, userId,status)
-    suspend fun getTasksByUserIdRemotely(taskId: Int) = remoteTaskRepository.getTasksByUserIdRemotely(taskId)
+    suspend fun getTasksByUserIdRemotely(taskId: Long) = remoteTaskRepository.getTasksByUserIdRemotely(taskId)
 }
